Refactor mui imports
