+++ components/profile/components/UserDetails/UserDetails.tsx
@@ -21,9 +21,9 @@ import { TelegramAccount } from 'pages/api/telegram/connect';
 import { shortenHex } from 'lib/utilities/strings';
 import useENSName from 'hooks/useENSName';
 import { UserAvatar } from 'lib/users/interfaces';
-import { useUserDetails } from 'components/profile/components/UserDetails/userDetails.hook';
 import { PublicUser } from 'pages/api/public/profile/[userPath]';
 import { hasNftAvatar } from 'lib/users/hasNftAvatar';
+import { useUserDetails } from 'components/profile/components/UserDetails/hooks/useUserDetails';
 import DescriptionModal from '../DescriptionModal';
 import UserPathModal from '../UserPathModal';
 import SocialModal from '../SocialModal';
