+++ components/profile/components/UserDetails/index.ts
@@ -1,0 +1,4 @@
+import UserDetails from './UserDetails';
+
+export default UserDetails;
+export * from './UserDetails';
