+++ components/profile/components/__tests__/UserDetails.test.tsx
@@ -4,7 +4,7 @@ import { UserDetails } from '@prisma/client';
 import { ThemeProvider } from '@mui/material/styles';
 import { createThemeLightSensitive } from 'theme';
 import IdentityModal from '../IdentityModal';
-import UserDetailsComponent from '../UserDetails/UserDetails';
+import UserDetailsComponent from '../UserDetails';
 
 const theme = createThemeLightSensitive('light');
