+++ components/profile/components/index.ts
@@ -1,4 +1,4 @@
 export { default as DescriptionModal } from './DescriptionModal';
 export { default as IdentityModal, getIdentityIcon, type IntegrationModel } from './IdentityModal';
 export { default as SocialModal } from './SocialModal';
-export { default as UserDetails } from './UserDetails/UserDetails';
+export { default as UserDetails } from './UserDetails';
